Improve to have more declarative code and move apply section and field related functions to its corresponding file
