+++ src/webapp/pages/form-page/disease-outbreak-event/utils/applyRulesInFormState.ts
@@ -1,9 +1,6 @@
 import { Rule } from "../../../../../domain/entities/Rule";
-import {
-    FormFieldState,
-    getFieldWithEmptyValue,
-} from "../../../../components/form/FormFieldsState";
-import { FormSectionState } from "../../../../components/form/FormSectionsState";
+import { FormFieldState } from "../../../../components/form/FormFieldsState";
+import { toggleSectionVisibilityByFieldValue } from "../../../../components/form/FormSectionsState";
 import { FormState } from "../../../../components/form/FormState";
 
 export function applyRulesInFormState(
@@ -30,58 +27,4 @@ export function applyRulesInFormState(
     }, currentFormState);
 
     return formStateWithRulesApplied;
-}
-
-function toggleSectionVisibilityByFieldValue(
-    section: FormSectionState,
-    fieldValue: FormFieldState["value"],
-    rule: Rule
-): FormSectionState {
-    if (rule.sectionIds.includes(section.id)) {
-        const subsections = section.subsections?.map(subsection => {
-            return toggleSectionVisibilityByFieldValue(subsection, fieldValue, rule);
-        });
-        return section.subsections
-            ? {
-                  ...section,
-                  subsections: subsections,
-                  isVisible: fieldValue === rule.fieldValue,
-                  fields:
-                      fieldValue === rule.fieldValue
-                          ? section.fields.map(field => ({
-                                ...field,
-                                isVisible: true,
-                            }))
-                          : hideFieldsAndSetToEmpty(section.fields),
-              }
-            : {
-                  ...section,
-                  isVisible: fieldValue === rule.fieldValue,
-                  fields:
-                      fieldValue === rule.fieldValue
-                          ? section.fields.map(field => ({
-                                ...field,
-                                isVisible: true,
-                            }))
-                          : hideFieldsAndSetToEmpty(section.fields),
-              };
-    }
-
-    return {
-        ...section,
-        subsections: section.subsections?.map(subsection =>
-            toggleSectionVisibilityByFieldValue(subsection, fieldValue, rule)
-        ),
-    };
-}
-
-function hideFieldsAndSetToEmpty(fields: FormFieldState[]): FormFieldState[] {
-    return fields.map(field => {
-        const fieldWithEmptyValue = getFieldWithEmptyValue(field);
-
-        return {
-            ...fieldWithEmptyValue,
-            isVisible: false,
-        };
-    });
 }
